refactor: move notification contracts to consensus model
