+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/junit/hedera/embedded/fakes/FakeNotificationEngine.java
@@ -3,16 +3,16 @@ package com.hedera.services.bdd.junit.hedera.embedded.fakes;
 
 import static java.util.Objects.requireNonNull;
 
-import com.swirlds.common.notification.Listener;
 import com.swirlds.common.notification.NotificationEngine;
 import com.swirlds.common.notification.NotificationResult;
-import com.swirlds.platform.listeners.PlatformStatusChangeListener;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.Future;
 import org.hiero.base.concurrent.futures.StandardFuture;
+import org.hiero.consensus.model.notification.Listener;
 import org.hiero.consensus.model.notification.Notification;
+import org.hiero.consensus.model.notification.PlatformStatusChangeListener;
 
 public class FakeNotificationEngine implements NotificationEngine {
     public final List<PlatformStatusChangeListener> statusChangeListeners = new CopyOnWriteArrayList<>();
