+++ platform-sdk/consensus-model/src/main/java/org/hiero/consensus/model/notification/PlatformStatusChangeListener.java
@@ -1,10 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.listeners;
-
-import com.swirlds.common.notification.DispatchMode;
-import com.swirlds.common.notification.DispatchModel;
-import com.swirlds.common.notification.DispatchOrder;
-import com.swirlds.common.notification.Listener;
+package org.hiero.consensus.model.notification;
 
 /**
  * Can be used to listen for changes in the platform status.
