+++ platform-sdk/consensus-otter-docker-app/src/testFixtures/java/module-info.java
@@ -2,7 +2,6 @@
 module org.hiero.consensus.otter.docker.app {
     requires transitive com.hedera.node.hapi;
     requires transitive com.swirlds.config.api;
-    requires transitive com.swirlds.platform.core;
     requires transitive org.hiero.consensus.metrics;
     requires transitive org.hiero.consensus.model;
     requires transitive org.hiero.otter.fixtures;
@@ -14,6 +13,7 @@ module org.hiero.consensus.otter.docker.app {
     requires com.swirlds.logging;
     requires com.swirlds.metrics.api;
     requires com.swirlds.platform.core.test.fixtures;
+    requires com.swirlds.platform.core;
     requires com.swirlds.state.api;
     requires com.swirlds.state.impl;
     requires com.swirlds.virtualmap;
