+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/notification/NotificationEngine.java
@@ -5,6 +5,8 @@ import com.swirlds.common.notification.internal.AsyncNotificationEngine;
 import java.util.concurrent.Future;
 import org.hiero.base.concurrent.futures.StandardFuture;
 import org.hiero.base.concurrent.manager.ThreadManager;
+import org.hiero.consensus.model.notification.DispatchMode;
+import org.hiero.consensus.model.notification.Listener;
 import org.hiero.consensus.model.notification.Notification;
 
 /**
