+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/notification/internal/AsyncNotificationEngine.java
@@ -1,9 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.swirlds.common.notification.internal;
 
-import com.swirlds.common.notification.DispatchMode;
-import com.swirlds.common.notification.DispatchOrder;
-import com.swirlds.common.notification.Listener;
 import com.swirlds.common.notification.NoListenersAvailableException;
 import com.swirlds.common.notification.NotificationResult;
 import java.util.Map;
@@ -12,6 +9,9 @@ import java.util.concurrent.Future;
 import java.util.function.Consumer;
 import org.hiero.base.concurrent.futures.StandardFuture;
 import org.hiero.base.concurrent.manager.ThreadManager;
+import org.hiero.consensus.model.notification.DispatchMode;
+import org.hiero.consensus.model.notification.DispatchOrder;
+import org.hiero.consensus.model.notification.Listener;
 import org.hiero.consensus.model.notification.Notification;
 
 public class AsyncNotificationEngine extends AbstractNotificationEngine {
