+++ platform-sdk/swirlds-common/src/timingSensitive/java/com/swirlds/common/notification/IntegerNotificationTests.java
@@ -35,6 +35,10 @@ import org.hiero.base.concurrent.framework.config.ThreadConfiguration;
 import org.hiero.base.concurrent.futures.ConcurrentFuturePool;
 import org.hiero.base.concurrent.futures.FuturePool;
 import org.hiero.base.concurrent.futures.StandardFuture;
+import org.hiero.consensus.model.notification.DispatchMode;
+import org.hiero.consensus.model.notification.DispatchModel;
+import org.hiero.consensus.model.notification.DispatchOrder;
+import org.hiero.consensus.model.notification.Listener;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
