+++ platform-sdk/swirlds-common/src/timingSensitive/java/com/swirlds/common/notification/SyncUnorderedIntegerListener.java
@@ -1,4 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.swirlds.common.notification;
 
+import org.hiero.consensus.model.notification.Listener;
+
 public interface SyncUnorderedIntegerListener extends Listener<IntegerNotification> {}
