+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/components/DefaultAppNotifier.java
@@ -2,8 +2,6 @@
 package com.swirlds.platform.components;
 
 import com.swirlds.common.notification.NotificationEngine;
-import com.swirlds.platform.listeners.PlatformStatusChangeListener;
-import com.swirlds.platform.listeners.PlatformStatusChangeNotification;
 import com.swirlds.platform.listeners.ReconnectCompleteListener;
 import com.swirlds.platform.listeners.ReconnectCompleteNotification;
 import com.swirlds.platform.listeners.StateWriteToDiskCompleteListener;
@@ -15,6 +13,8 @@ import com.swirlds.platform.system.state.notifications.StateHashedNotification;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import org.hiero.consensus.model.notification.IssNotification;
 import org.hiero.consensus.model.notification.IssNotification.IssType;
+import org.hiero.consensus.model.notification.PlatformStatusChangeListener;
+import org.hiero.consensus.model.notification.PlatformStatusChangeNotification;
 import org.hiero.consensus.model.status.PlatformStatus;
 
 /**
