+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/system/state/notifications/IssListener.java
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.swirlds.platform.system.state.notifications;
 
-import com.swirlds.common.notification.DispatchMode;
-import com.swirlds.common.notification.DispatchModel;
-import com.swirlds.common.notification.DispatchOrder;
-import com.swirlds.common.notification.Listener;
+import org.hiero.consensus.model.notification.DispatchMode;
+import org.hiero.consensus.model.notification.DispatchModel;
+import org.hiero.consensus.model.notification.DispatchOrder;
 import org.hiero.consensus.model.notification.IssNotification;
+import org.hiero.consensus.model.notification.Listener;
 
 /**
  * A method that listens for an ISS event. This listener provides no ordering guarantees with respect to
