+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/system/state/notifications/NewRecoveredStateListener.java
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.swirlds.platform.system.state.notifications;
 
-import com.swirlds.common.notification.DispatchMode;
-import com.swirlds.common.notification.DispatchModel;
-import com.swirlds.common.notification.DispatchOrder;
-import com.swirlds.common.notification.Listener;
+import org.hiero.consensus.model.notification.DispatchMode;
+import org.hiero.consensus.model.notification.DispatchModel;
+import org.hiero.consensus.model.notification.DispatchOrder;
+import org.hiero.consensus.model.notification.Listener;
 
 /**
  * A method that listens for the signed state created as a result of event recovery.
